requirements.txt
bcrypt==4.0.1
Flask==3.0.3
Flask-JWT-Extended==4.6.0
Flask-SQLAlchemy==3.1.1
PyJWT==2.8.0
passlib==1.7.4
python-dotenv==1.0.1
SQLAlchemy==2.0.30
